+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Full documentation for MIGraphX is available at
 
 * Added a dedicated logger for MIGraphX.
 * [Linux] Use HSA API to query number of chiplets for architectures where this is applicable (ex. gfx90a).
+* Added debug symbols for MIGraphX instructions such that parsed and compiled instructions can be tracked back to their ONNX origin node with MIGRAPHX_ENABLE_DEBUG_SYMBOLS (#4626)
 
 ### Changed